Config structure changed + moved mongodb config to database and added postgres for future implementation
