+++ src/core/domains/Database/interfaces/IDatabaseConfig.ts
@@ -1,0 +1,6 @@
+import IMongoDbConfig from "../mongodb/interfaces/IMongoDbConfig";
+
+export interface IDatabaseConfig {
+    mongodb: IMongoDbConfig;
+    postgres: null
+}
