+++ src/core/domains/Database/mongodb/providers/MongoDBProvider.ts
@@ -1,18 +1,18 @@
-import mongodbConfig from '@src/config/database/mongodb';
+import databaseConfig from '@src/config/database';
 import BaseProvider from "@src/core/base/Provider";
-import IMongoDbConfig from "@src/core/domains/database/mongodb/interfaces/IMongoDbConfig";
 import MongoDB from "@src/core/domains/database/mongodb/services/MongoDB";
 import { App } from "@src/core/services/App";
+import { IDatabaseConfig } from '../../interfaces/IDatabaseConfig';
 
 export default class MongoDBProvider extends BaseProvider
 {
-    protected config: IMongoDbConfig = mongodbConfig;
+    protected config: IDatabaseConfig = databaseConfig;
 
     public async register(): Promise<void>
     {
         this.log('Registering MongoDBProvider');
         
-        App.setContainer('mongodb', new MongoDB(this.config))
+        App.setContainer('mongodb', new MongoDB(this.config.mongodb))
     }
 
     public async boot(): Promise<void>
