+++ src/core/domains/express/providers/RoutesProvider.ts
@@ -1,4 +1,4 @@
-import expressConfig from '@src/config/http/express';
+import httpConfig from '@src/config/http';
 import BaseProvider from "@src/core/base/Provider";
 import IExpressConfig from '@src/core/domains/express/interfaces/IExpressConfig';
 import ExpressProvider from "@src/core/domains/express/providers/ExpressProvider";
@@ -8,7 +8,7 @@ import { App } from '@src/core/services/App';
 
 export default class RoutesProvider extends BaseProvider
 {
-    protected config: IExpressConfig = expressConfig;
+    protected config: IExpressConfig = httpConfig;
     
     public async register(): Promise<void> {
         if(!this.config.enabled) {
